+++ desdeo/api/models/gdm_aggregate.py
@@ -1,61 +1,18 @@
-"""Classes for group decision making"""
+"""Classes for group decision making, aggregating all different types of data classes 
+from across the derived classes of BaseSolverResult
+"""
 
 import json
-from typing import Literal
 
 from sqlalchemy.types import TypeDecorator
 from sqlmodel import SQLModel, Field, Relationship, JSON, Column
-from pydantic import ValidationError
 
-from desdeo.api.models import ReferencePoint
+from desdeo.api.models import BasePreferenceResults
+from desdeo.api.models.gnimbus import (
+    VotingPreferenceResults,
+    GNIMBUSPreferenceResults
+)
 from desdeo.tools import SolverResults
-
-class SolverResultType(TypeDecorator):
-    """Dunno why the solver results wouldn't serialize but this should do the trick"""
-    impl = JSON
-
-    def process_bind_param(self, value, dialect):
-        if isinstance(value, list):
-            solver_list = []
-            for item in value:
-                if isinstance(item, SolverResults):
-                    solver_list.append(item.model_dump_json())
-            return json.dumps(solver_list)
-    
-    def process_result_value(self, value, dialect):
-        if value is not None:
-            value_list = json.loads(value)
-            solver_list: SolverResults = []
-            for item in value_list:
-                item_dict = json.loads(item)
-                try:
-                    solver_list.append(SolverResults.model_validate(item_dict))
-                except ValidationError as e:
-                    print(f"Validation error when deserializing SolverResults: {e}")
-                    continue
-            return solver_list
-        
-class ReferencePointDictType(TypeDecorator):
-    """A converter for dict of int and preferences"""
-    impl = JSON
-
-    def process_bind_param(self, value, dialect):
-        if isinstance(value, dict):
-            for key, item in value.items():
-                if isinstance(item, ReferencePoint):
-                    value[key] = item.model_dump_json()
-            return json.dumps(value)
-    
-    def process_result_value(self, value, dialect):
-        dictionary = json.loads(value)
-        for key, item in dictionary.items():
-            if item == None:
-                print("Something's wrong... Database has a NoneType entry.")
-            try:
-                dictionary[key] = ReferencePoint.model_validate(json.loads(item))
-            except ValidationError as e:
-                print(f"Validation error when deserializing PreferencePoint: {e}")
-        return dictionary
 
 class PreferenceResultType(TypeDecorator):
     """A converter of Preference/Result types"""
@@ -77,9 +34,10 @@ class PreferenceResultType(TypeDecorator):
                 case "voting":
                     valdeser = VotingPreferenceResults.model_validate(jsoned)
                     return valdeser
-                case "nimbus":
-                    valdeser = NIMBUSPreferenceResults.model_validate(jsoned)
+                case "gnimbus":
+                    valdeser = GNIMBUSPreferenceResults.model_validate(jsoned)
                     return valdeser
+                # As the different methods are implemented, add new types
                 case _:
                     return None
         return None
@@ -106,22 +64,6 @@ class GroupPublic(GroupBase):
     owner_id: int
     user_ids: list[int]
     problem_id: int
-
-class BasePreferenceResults(SQLModel):
-    """A base class for a method specific preference and results"""
-    method: str = "unset"
-
-class NIMBUSPreferenceResults(BasePreferenceResults):
-    """A NIMBUS preference and result class"""
-    method: str = "nimbus"
-    set_preferences: dict[int, ReferencePoint] = Field(sa_column=Column(ReferencePointDictType))
-    results: list[SolverResults] = Field(sa_column=Column(SolverResultType))
-
-class VotingPreferenceResults(BasePreferenceResults):
-    """A voting preferences and results"""
-    method: str = "voting"
-    set_preferences: dict[int, int] = Field() # A user votes for an index from the results
-    results: SolverResults = Field() # The winning result
 
 class GroupIteration(SQLModel, table=True):
     """Table model for Group Iteration (we could extend this in various ways)"""
